# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aboubakry Dieng Ges-Compte API",
        "description": "Documentation de l'API Ges-Compte développée par Aboubakry Dieng",
        "contact": {
            "name": "Aboubakry Dieng",
            "email": "aboubakry.dieng@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gescompte-1.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister tous les comptes",
                "operationId": "e08298dbea7e337b2ce3af7d443388c7",
                "responses": {
                    "200": {
                        "description": "Liste des comptes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte",
                "operationId": "c4b9822f8407aeeaa6ed3cf36ed4ad35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "client_id"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "epargne",
                                            "cheque",
                                            "courant"
                                        ]
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "bloque",
                                            "ferme"
                                        ]
                                    },
                                    "client_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comptes/{id}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Obtenir un compte spécifique",
                "operationId": "78a5bf0e4067fef67a2df695f8e865a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Mettre à jour un compte",
                "operationId": "e3881fc7f13fec9e0d43777cb4392511",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "epargne",
                                            "cheque",
                                            "courant"
                                        ]
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "bloque",
                                            "ferme"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Compte"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Supprimer un compte",
                "operationId": "9d64c974a58e7a3ee108815cd435c1c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Compte supprimé"
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Compte": {
                "title": "Compte",
                "description": "Représente un compte bancaire",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "numero": {
                        "type": "string",
                        "example": "CMPT2025A1B2C3D4"
                    },
                    "type": {
                        "type": "string",
                        "example": "epargne"
                    },
                    "statut": {
                        "type": "string",
                        "example": "actif"
                    },
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}